+++ doc/version_history.md
@@ -8,12 +8,15 @@
 >     - $versions.spring_cloud replace to ${springCloudVersion}
 >     - $versions.alibaba_cloud replace to ${springCloudAlibabaVersion}
 >   - [Upgrade] 优化junit测试代码
-> - acp-core
->   - [Upgrade] bouncycastle 的组件由15on替换为15to18后缀
 > - acp-spring-boot-starter
 >   - [Upgrade] controller切面中，google ImmutableList替换为Kotlin的MutableList
 > - acp-spring-cloud-starter
 >   - [Upgrade] AcpCloudResourceServerAutoConfiguration 中关闭 formLogin 登录界面
+>   - [Upgrade] 增加依赖 org.springframework.cloud:spring-cloud-starter-bootstrap
+>   - [Upgrade] 去除ribbon依赖
+>   - [Upgrade] 更换security依赖
+>   - [Upgrade] kafka stream 使用函数式风格改写
+>   - [Upgrade] 增加 AcpCloudFeignAutoConfiguration，重写 SentinelFeign 和 SentinelInvocationHandler，以适配 Spring Cloud 2020.0.0
 > - test
 >   - [Upgrade] logback 配置属性重命名
 >     - logging.pattern.rolling-file-name → logging.logback.rollingpolicy.file-name-pattern
@@ -21,10 +24,14 @@
 >     - logging.file.max-size → logging.logback.rollingpolicy.max-file-size
 >     - logging.file.total-size-cap → logging.logback.rollingpolicy.total-size-cap
 >     - logging.file.max-history → logging.logback.rollingpolicy.max-history
+>   - [Upgrade] 修改feign FallbackFactory，不再使用netflix-hystrix
+>   [Upgrade] admin-server、gateway-server 去除 ribbon 依赖
 > - [Upgrade] 升级依赖
 >   - kotlin coroutines 1.4.2
 >   - kotlin 1.4.21
 >   - Spring Boot 2.4.1
+>   - Spring Cloud 2020.0.0
+>   - Spring Boot Admin 2.3.1
 >   - knife4j 3.0.2
 >   - jupiter 5.7.0
 >   - junit-platform 1.7.0
@@ -37,7 +44,6 @@
 >   - commons-codec 1.15
 >   - commons-text 1.9
 >   - commons-lang3 3.11
->   - bouncycastle 1.67
 >   - xstream 1.4.14
 >   - jackson 2.11.3
 >   - zip4j 2.6.4